\name{WriteScoreR-package}
\alias{WriteScoreR-package}
\alias{WriteScoreR}
\docType{package}
\title{
Utilities for creating questions for the ScoreR online grading system.
}
\description{
The 
}
\details{
\tabular{ll}{
Package: \tab WriteScoreR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-05-21\cr
License: \tab What license is it under?\cr
}
We'll say something here.
}
\author{
Danny Kaplan

Maintainer: Danny Kaplan <dtkaplan@gmail.com>
}
\references{
The references here.
}

\keyword{ package }
\seealso{
\code{\link[mosaic:mosaic-package]{mosaic}}
}
\examples{
sin(3)
}
